# Common.Contracts/openapi_v3.yaml
openapi: 3.0.4
info:
  title: LPshows.live API v3
  description: This is the API for the Linkin Park Concert Calendar fan project
  version: 3.0.0
servers:
  - url: /
  - url: http://localhost:5133
paths:
...
  /v3/Venues:
    post:
      tags:
        - Venues
      summary: Creates a new venue
      operationId: CreateVenue
      requestBody:
        description: data of the new venue
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateVenueRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateVenueRequest"
        required: true
      responses:
        "200":
          description: ""
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues
      security:
        - Bearer: []
    get:
      tags:
        - Venues
      summary: Returns an unfiltered list of all venues
      operationId: GetAllVenues
      responses:
        "200":
          description: list of all venues
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues
      security:
        - Bearer: []
  /v3/Venues/{venueId}:
    put:
      tags:
        - Venues
      summary: Updates the information about a venue. Please note that updates of the
        name have to be made through special routes.
      operationId: UpdateVenue
      parameters:
        - name: venueId
          in: path
          description: ID of the venue
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      requestBody:
        description: New information of the venue. Partial updates are not possible.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateVenueRequest"
        required: true
      responses:
        "200":
          description: all information about the venue
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/VenueWithDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/VenueWithDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/VenueWithDetails"
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: PUT
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues/{venueId}
        requestParameters:
          integration.request.path.venueId: method.request.path.venueId
      security:
        - Bearer: []
    get:
      tags:
        - Venues
      summary: Returns the basic information about a venue by its ID
      operationId: GetVenueById
      parameters:
        - name: venueId
          in: path
          description: ID of the venue
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      responses:
        "200":
          description: basic information about the venue
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/Venue"
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
            text/json:
              schema:
                $ref: "#/components/schemas/Venue"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues/{venueId}
        requestParameters:
          integration.request.path.venueId: method.request.path.venueId
    delete:
      tags:
        - Venues
      summary: Deletes a venue
      operationId: DeleteVenueById
      parameters:
        - name: venueId
          in: path
          description: ID of the venue
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: DELETE
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues/{venueId}
        requestParameters:
          integration.request.path.venueId: method.request.path.venueId
      security:
        - Bearer: []
  /v3/Venues/{venueId}/details:
    get:
      tags:
        - Venues
      summary: Returns the detailed information about a venue by its ID including
        information about the city and country
      operationId: GetVenueWithCityById
      parameters:
        - name: venueId
          in: path
          description: ID of the venue
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      responses:
        "200":
          description: detailed information about the venue
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/VenueWithCity"
            application/json:
              schema:
                $ref: "#/components/schemas/VenueWithCity"
            text/json:
              schema:
                $ref: "#/components/schemas/VenueWithCity"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues/{venueId}/details
        requestParameters:
          integration.request.path.venueId: method.request.path.venueId
  /v3/Venues/{venueId}/names:
    post:
      tags:
        - Venues
      summary: Adds a new name to a venue for a given time range
      operationId: AddNewVenueName
      parameters:
        - name: venueId
          in: path
          description: ID of the venue
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddVenueNameRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/AddVenueNameRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/AddVenueNameRequest"
        required: true
      responses:
        "200":
          description: Venue information with details
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/VenueWithDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/VenueWithDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/VenueWithDetails"
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues/{venueId}/names
        requestParameters:
          integration.request.path.venueId: method.request.path.venueId
      security:
        - Bearer: []
  /v3/Venues/{venueId}/names/{venueNameId}:
    put:
      tags:
        - Venues
      summary: Changes an existing name of a venue
      operationId: UpdateVenueName
      parameters:
        - name: venueId
          in: path
          description: ID of the venue
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
        - name: venueNameId
          in: path
          description: ID of the venue name
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueNameRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueNameRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateVenueNameRequest"
        required: true
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: PUT
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues/{venueId}/names/{venueNameId}
        requestParameters:
          integration.request.path.venueId: method.request.path.venueId
          integration.request.path.venueNameId: method.request.path.venueNameId
      security:
        - Bearer: []
    delete:
      tags:
        - Venues
      summary: Deletes a previous name of a venue
      operationId: DeleteVenueName
      parameters:
        - name: venueId
          in: path
          description: ID of the venue
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
        - name: venueNameId
          in: path
          description: ID of the venue name
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: DELETE
        uri: https://api.dev.lpshows.felixsfd.de/v3/venues/{venueId}/names/{venueNameId}
        requestParameters:
          integration.request.path.venueId: method.request.path.venueId
          integration.request.path.venueNameId: method.request.path.venueNameId
      security:
        - Bearer: []
components:
  schemas:
...
    AddVenueNameRequest:
      required:
        - name
      type: object
      properties:
        name:
          type: string
        from:
          type: string
          format: date
        to:
          type: string
          format: date
    CityWithCountry:
      required:
        - country
        - countryCode
        - name
        - nativeName
      type: object
      properties:
        country:
          $ref: "#/components/schemas/Country"
        state:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/State"
        countryCode:
          type: string
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        stateCode:
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    Country:
      required:
        - isoCode
        - name
        - nativeName
      type: object
      properties:
        isoCode:
          maxLength: 3
          minLength: 3
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    CreateVenueRequest:
      required:
        - countryCode
        - currentName
        - timeZone
      type: object
      properties:
        countryCode:
          maxLength: 3
          minLength: 3
          type: string
        stateCode:
          maxLength: 3
          type: string
        cityId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        currentName:
          type: string
        timeZone:
          type: string
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          format: double
    PreviousVenueName:
      required:
        - venueId
        - id
        - name
        - usedFrom
      type: object
      properties:
        venueId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        name:
          type: string
        usedFrom:
          type: string
          format: date
        usedUntil:
          type: string
          format: date
...
    State:
      required:
        - countryCode
        - code
        - name
        - nativeName
      type: object
      properties:
        countryCode:
          type: string
        code:
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    UpdateVenueNameRequest:
      required:
        - name
      type: object
      properties:
        name:
          type: string
        from:
          type: string
          format: date
        to:
          type: string
          nullable: true
          format: date
    UpdateVenueRequest:
      required:
        - countryCode
        - timeZone
      type: object
      properties:
        countryCode:
          maxLength: 3
          minLength: 3
          type: string
        stateCode:
          maxLength: 3
          type: string
          nullable: true
        cityId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        timeZone:
          type: string
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
    Venue:
      required:
        - countryCode
        - currentName
        - timeZone
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        countryCode:
          maxLength: 3
          minLength: 3
          type: string
        stateCode:
          maxLength: 3
          type: string
        cityId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        currentName:
          type: string
        timeZone:
          type: string
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
    VenueWithCity:
      required:
        - city
        - countryCode
        - currentName
        - timeZone
      type: object
      properties:
        city:
          $ref: "#/components/schemas/CityWithCountry"
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        countryCode:
          maxLength: 3
          minLength: 3
          type: string
        stateCode:
          maxLength: 3
          type: string
          nullable: true
        cityId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        currentName:
          type: string
        timeZone:
          type: string
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
    VenueWithDetails:
      required:
        - venueNames
        - city
        - countryCode
        - currentName
        - timeZone
      type: object
      properties:
        venueNames:
          type: array
          items:
            $ref: "#/components/schemas/PreviousVenueName"
        city:
          $ref: "#/components/schemas/CityWithCountry"
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        countryCode:
          maxLength: 3
          minLength: 3
          type: string
        stateCode:
          maxLength: 3
          type: string
          nullable: true
        cityId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        currentName:
          type: string
        timeZone:
          type: string
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
    
    # Schemas from v2
...
  securitySchemes:
    Bearer:
      type: http
      description: JWT Bearer authentication
      scheme: Bearer
    ConcertsCognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        type: cognito_user_pools
        providerARNs:
          - Fn::ImportValue:
              Fn::Sub: '${BaseProductName}-cognito-authorizer-user-pool'
...